switched the .local and regular story file for jobsPages
